+++ CumulusMX/ExtensionMethods/InputExtensions.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace CumulusMX.ExtensionMethods
+{
+    public static class InputExtensions
+    {
+        public static int LimitToRange(
+            this int value, int inclusiveMinimum, int inclusiveMaximum)
+        {
+            if (value < inclusiveMinimum) { return inclusiveMinimum; }
+            if (value > inclusiveMaximum) { return inclusiveMaximum; }
+            return value;
+        }
+    }
+}
