--- CumulusMX/Extensions/TwitterConnector.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Cumulus4.Extensions
-{
-    public class TwitterConnector : IDataReporter
-    {
-    }
-}
