+++ CumulusMX/Web/Controllers/SettingsController.cs
@@ -3,12 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Web;
-using Cumulus4.Data;
+using CumulusMX.Data;
 using Unosquare.Labs.EmbedIO;
 using Unosquare.Labs.EmbedIO.Constants;
 using Unosquare.Labs.EmbedIO.Modules;
 
-namespace Cumulus4.Web.Controllers
+namespace CumulusMX.Web.Controllers
 {
     public class SettingsController : ControllerBase
     {
